+++ internal/services/servicebus/servicebus_queue_resource_test.go
@@ -6,10 +6,10 @@ import (
 	"regexp"
 	"testing"
 
+	"github.com/hashicorp/go-azure-sdk/resource-manager/servicebus/2021-06-01-preview/queues"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance/check"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/servicebus/parse"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/utils"
 )
@@ -60,6 +60,7 @@ func TestAccServiceBusQueue_update(t *testing.T) {
 				check.That(data.ResourceName).Key("enable_batched_operations").HasValue("true"),
 			),
 		},
+		data.ImportStep(),
 		{
 			Config: r.update(data),
 			Check: acceptance.ComposeTestCheckFunc(
@@ -130,7 +131,7 @@ func TestAccServiceBusQueue_enablePartitioningForPremiumError(t *testing.T) {
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
 			Config:      r.enablePartitioningForPremiumError(data),
-			ExpectError: regexp.MustCompile("Partitioning Entities is not supported in Premium SKU and must be disabled"),
+			ExpectError: regexp.MustCompile("partitioning Entities is not supported in Premium SKU and must be disabled"),
 		},
 	})
 }
@@ -366,17 +367,17 @@ func TestAccServiceBusQueue_status(t *testing.T) {
 }
 
 func (t ServiceBusQueueResource) Exists(ctx context.Context, clients *clients.Client, state *pluginsdk.InstanceState) (*bool, error) {
-	id, err := parse.QueueID(state.ID)
+	id, err := queues.ParseQueueID(state.ID)
 	if err != nil {
 		return nil, err
 	}
 
-	resp, err := clients.ServiceBus.QueuesClient.Get(ctx, id.ResourceGroup, id.NamespaceName, id.Name)
+	resp, err := clients.ServiceBus.QueuesClient.Get(ctx, *id)
 	if err != nil {
-		return nil, fmt.Errorf("reading Service Bus NameSpace Queue (%s): %+v", id.String(), err)
+		return nil, fmt.Errorf("retrieving %s: %+v", *id, err)
 	}
 
-	return utils.Bool(resp.ID != nil), nil
+	return utils.Bool(resp.Model != nil), nil
 }
 
 func (ServiceBusQueueResource) basic(data acceptance.TestData) string {