+++ internal/services/servicebus/servicebus_topic_resource.go
@@ -5,11 +5,12 @@ import (
 	"log"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/preview/servicebus/mgmt/2021-06-01-preview/servicebus"
+	"github.com/hashicorp/go-azure-helpers/lang/response"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/servicebus/2021-06-01-preview/namespaces"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/servicebus/2021-06-01-preview/topics"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/servicebus/parse"
 	azValidate "github.com/hashicorp/terraform-provider-azurerm/internal/services/servicebus/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
@@ -25,7 +26,7 @@ func resourceServiceBusTopic() *pluginsdk.Resource {
 		Delete: resourceServiceBusTopicDelete,
 
 		Importer: pluginsdk.ImporterValidatingResourceId(func(id string) error {
-			_, err := parse.TopicID(id)
+			_, err := topics.ParseTopicID(id)
 			return err
 		}),
 
@@ -54,16 +55,16 @@ func resourceServiceBusTopicSchema() map[string]*pluginsdk.Schema {
 			Type:         pluginsdk.TypeString,
 			Required:     true,
 			ForceNew:     true,
-			ValidateFunc: azValidate.NamespaceID,
+			ValidateFunc: namespaces.ValidateNamespaceID,
 		},
 
 		"status": {
 			Type:     pluginsdk.TypeString,
 			Optional: true,
-			Default:  string(servicebus.EntityStatusActive),
+			Default:  string(topics.EntityStatusActive),
 			ValidateFunc: validation.StringInSlice([]string{
-				string(servicebus.EntityStatusActive),
-				string(servicebus.EntityStatusDisabled),
+				string(topics.EntityStatusActive),
+				string(topics.EntityStatusDisabled),
 			}, false),
 		},
 
@@ -140,79 +141,78 @@ func resourceServiceBusTopicCreateUpdate(d *pluginsdk.ResourceData, meta interfa
 	defer cancel()
 	log.Printf("[INFO] preparing arguments for Azure ServiceBus Topic creation.")
 
-	status := d.Get("status").(string)
-
-	enableBatchedOps := d.Get("enable_batched_operations").(bool)
-	enableExpress := d.Get("enable_express").(bool)
-	enablePartitioning := d.Get("enable_partitioning").(bool)
-	maxSize := int32(d.Get("max_size_in_megabytes").(int))
-	requiresDuplicateDetection := d.Get("requires_duplicate_detection").(bool)
-	supportOrdering := d.Get("support_ordering").(bool)
-
-	var resourceId parse.TopicId
+	var id topics.TopicId
 	if namespaceIdLit := d.Get("namespace_id").(string); namespaceIdLit != "" {
-		namespaceId, _ := parse.NamespaceID(namespaceIdLit)
-		resourceId = parse.NewTopicID(namespaceId.SubscriptionId, namespaceId.ResourceGroup, namespaceId.Name, d.Get("name").(string))
+		namespaceId, err := topics.ParseNamespaceID(namespaceIdLit)
+		if err != nil {
+			return err
+		}
+		id = topics.NewTopicID(namespaceId.SubscriptionId, namespaceId.ResourceGroupName, namespaceId.NamespaceName, d.Get("name").(string))
 	}
 
 	if d.IsNewResource() {
-		existing, err := client.Get(ctx, resourceId.ResourceGroup, resourceId.NamespaceName, resourceId.Name)
+		existing, err := client.Get(ctx, id)
 		if err != nil {
-			if !utils.ResponseWasNotFound(existing.Response) {
-				return fmt.Errorf("checking for presence of existing %s: %+v", resourceId, err)
+			if !response.WasNotFound(existing.HttpResponse) {
+				return fmt.Errorf("checking for presence of existing %s: %+v", id, err)
 			}
 		}
 
-		if !utils.ResponseWasNotFound(existing.Response) {
-			return tf.ImportAsExistsError("azurerm_service_fabric_cluster", resourceId.ID())
+		if !response.WasNotFound(existing.HttpResponse) {
+			return tf.ImportAsExistsError("azurerm_service_fabric_cluster", id.ID())
 		}
 	}
 
-	parameters := servicebus.SBTopic{
-		Name: utils.String(resourceId.Name),
-		SBTopicProperties: &servicebus.SBTopicProperties{
-			Status:                     servicebus.EntityStatus(status),
-			EnableBatchedOperations:    utils.Bool(enableBatchedOps),
-			EnableExpress:              utils.Bool(enableExpress),
-			EnablePartitioning:         utils.Bool(enablePartitioning),
-			MaxSizeInMegabytes:         utils.Int32(maxSize),
-			RequiresDuplicateDetection: utils.Bool(requiresDuplicateDetection),
-			SupportOrdering:            utils.Bool(supportOrdering),
+	status := topics.EntityStatus(d.Get("status").(string))
+	parameters := topics.SBTopic{
+		Name: utils.String(id.TopicName),
+		Properties: &topics.SBTopicProperties{
+			Status:                     &status,
+			EnableBatchedOperations:    utils.Bool(d.Get("enable_batched_operations").(bool)),
+			EnableExpress:              utils.Bool(d.Get("enable_express").(bool)),
+			EnablePartitioning:         utils.Bool(d.Get("enable_partitioning").(bool)),
+			MaxSizeInMegabytes:         utils.Int64(int64(d.Get("max_size_in_megabytes").(int))),
+			RequiresDuplicateDetection: utils.Bool(d.Get("requires_duplicate_detection").(bool)),
+			SupportOrdering:            utils.Bool(d.Get("support_ordering").(bool)),
 		},
 	}
 
 	if autoDeleteOnIdle := d.Get("auto_delete_on_idle").(string); autoDeleteOnIdle != "" {
-		parameters.SBTopicProperties.AutoDeleteOnIdle = utils.String(autoDeleteOnIdle)
+		parameters.Properties.AutoDeleteOnIdle = utils.String(autoDeleteOnIdle)
 	}
 
 	if defaultTTL := d.Get("default_message_ttl").(string); defaultTTL != "" {
-		parameters.SBTopicProperties.DefaultMessageTimeToLive = utils.String(defaultTTL)
+		parameters.Properties.DefaultMessageTimeToLive = utils.String(defaultTTL)
 	}
 
 	if duplicateWindow := d.Get("duplicate_detection_history_time_window").(string); duplicateWindow != "" {
-		parameters.SBTopicProperties.DuplicateDetectionHistoryTimeWindow = utils.String(duplicateWindow)
+		parameters.Properties.DuplicateDetectionHistoryTimeWindow = utils.String(duplicateWindow)
 	}
 
 	// We need to retrieve the namespace because Premium namespace works differently from Basic and Standard
 	namespacesClient := meta.(*clients.Client).ServiceBus.NamespacesClient
-	namespace, err := namespacesClient.Get(ctx, resourceId.ResourceGroup, resourceId.NamespaceName)
+	namespaceId := namespaces.NewNamespaceID(id.SubscriptionId, id.ResourceGroupName, id.NamespaceName)
+	resp, err := namespacesClient.Get(ctx, namespaceId)
 	if err != nil {
-		return fmt.Errorf("retrieving ServiceBus Namespace %q (Resource Group %q): %+v", resourceId.NamespaceName, resourceId.ResourceGroup, err)
+		return fmt.Errorf("retrieving ServiceBus Namespace %q (Resource Group %q): %+v", id.NamespaceName, id.ResourceGroupName, err)
 	}
 
 	// output of `max_message_size_in_kilobytes` is also set in non-Premium namespaces, with a value of 256
 	if v, ok := d.GetOk("max_message_size_in_kilobytes"); ok && v.(int) != 256 {
-		if namespace.Sku.Name != servicebus.SkuNamePremium {
-			return fmt.Errorf("ServiceBus Topic %q does not support input on `max_message_size_in_kilobytes` in %s SKU and should be removed", resourceId.Name, namespace.Sku.Name)
+		if model := resp.Model; model != nil {
+			if model.Sku.Name != namespaces.SkuNamePremium {
+				return fmt.Errorf("%s does not support input on `max_message_size_in_kilobytes` in %s SKU and should be removed", id, model.Sku.Name)
+			}
+			parameters.Properties.MaxMessageSizeInKilobytes = utils.Int64(int64(v.(int)))
 		}
-		parameters.SBTopicProperties.MaxMessageSizeInKilobytes = utils.Int64(int64(v.(int)))
+
 	}
 
-	if _, err := client.CreateOrUpdate(ctx, resourceId.ResourceGroup, resourceId.NamespaceName, resourceId.Name, parameters); err != nil {
-		return fmt.Errorf("creating/updating %s: %v", resourceId, err)
+	if _, err := client.CreateOrUpdate(ctx, id, parameters); err != nil {
+		return fmt.Errorf("creating/updating %s: %v", id, err)
 	}
 
-	d.SetId(resourceId.ID())
+	d.SetId(id.ID())
 	return resourceServiceBusTopicRead(d, meta)
 }
 
@@ -221,58 +221,67 @@ func resourceServiceBusTopicRead(d *pluginsdk.ResourceData, meta interface{}) er
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.TopicID(d.Id())
+	id, err := topics.ParseTopicID(d.Id())
 	if err != nil {
 		return err
 	}
 
-	resp, err := client.Get(ctx, id.ResourceGroup, id.NamespaceName, id.Name)
+	resp, err := client.Get(ctx, *id)
 	if err != nil {
-		if utils.ResponseWasNotFound(resp.Response) {
+		if response.WasNotFound(resp.HttpResponse) {
 			d.SetId("")
 			return nil
 		}
 		return fmt.Errorf("retrieving %s: %+v", id, err)
 	}
 
-	d.Set("name", id.Name)
-	d.Set("namespace_id", parse.NewNamespaceID(id.SubscriptionId, id.ResourceGroup, id.NamespaceName).ID())
+	d.Set("name", id.TopicName)
+	d.Set("namespace_id", topics.NewNamespaceID(id.SubscriptionId, id.ResourceGroupName, id.NamespaceName).ID())
 
-	if props := resp.SBTopicProperties; props != nil {
-		d.Set("status", string(props.Status))
-		d.Set("auto_delete_on_idle", props.AutoDeleteOnIdle)
-		d.Set("default_message_ttl", props.DefaultMessageTimeToLive)
+	if model := resp.Model; model != nil {
+		if props := model.Properties; props != nil {
+			status := ""
+			if v := props.Status; v != nil {
+				status = string(*v)
+			}
+			d.Set("status", status)
+			d.Set("auto_delete_on_idle", props.AutoDeleteOnIdle)
+			d.Set("default_message_ttl", props.DefaultMessageTimeToLive)
 
-		if window := props.DuplicateDetectionHistoryTimeWindow; window != nil && *window != "" {
-			d.Set("duplicate_detection_history_time_window", window)
-		}
-
-		d.Set("enable_batched_operations", props.EnableBatchedOperations)
-		d.Set("enable_express", props.EnableExpress)
-		d.Set("enable_partitioning", props.EnablePartitioning)
-		d.Set("max_message_size_in_kilobytes", props.MaxMessageSizeInKilobytes)
-		d.Set("requires_duplicate_detection", props.RequiresDuplicateDetection)
-		d.Set("support_ordering", props.SupportOrdering)
-
-		if maxSizeMB := props.MaxSizeInMegabytes; maxSizeMB != nil {
-			maxSize := int(*props.MaxSizeInMegabytes)
-
-			// if the topic is in a premium namespace and partitioning is enabled then the
-			// max size returned by the API will be 16 times greater than the value set
-			if partitioning := props.EnablePartitioning; partitioning != nil && *partitioning {
-				namespacesClient := meta.(*clients.Client).ServiceBus.NamespacesClient
-				namespace, err := namespacesClient.Get(ctx, id.ResourceGroup, id.NamespaceName)
-				if err != nil {
-					return err
-				}
-
-				if namespace.Sku.Name != servicebus.SkuNamePremium {
-					const partitionCount = 16
-					maxSize = int(*props.MaxSizeInMegabytes / partitionCount)
-				}
+			if window := props.DuplicateDetectionHistoryTimeWindow; window != nil && *window != "" {
+				d.Set("duplicate_detection_history_time_window", window)
 			}
 
-			d.Set("max_size_in_megabytes", maxSize)
+			d.Set("enable_batched_operations", props.EnableBatchedOperations)
+			d.Set("enable_express", props.EnableExpress)
+			d.Set("enable_partitioning", props.EnablePartitioning)
+			d.Set("max_message_size_in_kilobytes", props.MaxMessageSizeInKilobytes)
+			d.Set("requires_duplicate_detection", props.RequiresDuplicateDetection)
+			d.Set("support_ordering", props.SupportOrdering)
+
+			if maxSizeMB := props.MaxSizeInMegabytes; maxSizeMB != nil {
+				maxSize := int(*props.MaxSizeInMegabytes)
+
+				// if the topic is in a premium namespace and partitioning is enabled then the
+				// max size returned by the API will be 16 times greater than the value set
+				if partitioning := props.EnablePartitioning; partitioning != nil && *partitioning {
+					namespacesClient := meta.(*clients.Client).ServiceBus.NamespacesClient
+					namespaceId := namespaces.NewNamespaceID(id.SubscriptionId, id.ResourceGroupName, id.NamespaceName)
+					namespaceResp, err := namespacesClient.Get(ctx, namespaceId)
+					if err != nil {
+						return err
+					}
+
+					if namespaceModel := namespaceResp.Model; namespaceModel != nil {
+						if namespaceModel.Sku.Name != namespaces.SkuNamePremium {
+							const partitionCount = 16
+							maxSize = int(*props.MaxSizeInMegabytes / partitionCount)
+						}
+					}
+				}
+
+				d.Set("max_size_in_megabytes", maxSize)
+			}
 		}
 	}
 
@@ -284,14 +293,14 @@ func resourceServiceBusTopicDelete(d *pluginsdk.ResourceData, meta interface{}) 
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.TopicID(d.Id())
+	id, err := topics.ParseTopicID(d.Id())
 	if err != nil {
 		return err
 	}
 
-	resp, err := client.Delete(ctx, id.ResourceGroup, id.NamespaceName, id.Name)
+	resp, err := client.Delete(ctx, *id)
 	if err != nil {
-		if !utils.ResponseWasNotFound(resp) {
+		if !response.WasNotFound(resp.HttpResponse) {
 			return fmt.Errorf("deleting %s: %+v", id, err)
 		}
 	}